Homework 2023-06-13

+++ src/main/java/uz/najottalim/bankingapp/dto/RoleDto.java
@@ -4,12 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class RoleDto {
-    private Integer id;
+    private Long id;
     private String name;
 }
